+++ core/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 1.2.0
+## [1.2.0](https://github.com/availproject/avail-light/tree/avail-light-core-v1.2.0) - 2025-02-06
 
 - Remove multiaddress metric attribute
 - Add timestamp to v2 header API